pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ridestream"
version = "2.0.0"
description = "Real-time ride streaming data pipeline"
authors = [{name = "Engineering Team", email = "eng@ridestream.dev"}]
requires-python = ">=3.12"
license = {text = "Apache-2.0"}

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "kafka-python>=2.0.2",
    "boto3>=1.28.0",
    "minio>=7.2.0",
    "dbt-core>=1.7.0",
    "dbt-trino>=1.7.0",
    "sqlalchemy>=2.0",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "trino>=0.337.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-docker>=2.0.0",
    "testcontainers[kafka]>=3.7.0",
    "mypy>=1.5.0",
    "mypy-boto3-s3>=1.28.0",
    "mypy-boto3-kinesis>=1.28.0",
    "ruff>=0.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "bandit>=1.7.0",
    "pre-commit>=3.4.0",
    "pre-commit-hooks>=4.4.0",
    "detect-secrets>=1.4.0",
    "pyparsing>=3.0.0",
    "moto[s3,glue,kinesis,kms,secretsmanager]>=4.2.0",
    "docker>=6.0.0",
]

[project.scripts]
ridestream = "ridestream.entry_points.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ridestream"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
addopts = "--strict-markers --strict-config"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "smoke: smoke tests",
    "aws: requires AWS credentials (skips without AWS_ACCOUNT_ID)",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "kafka.*",
    "minio.*",
    "testcontainers.*",
    "boto3.*",
    "botocore.*",
    "trino.*",
    "moto.*",
    "pyspark.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ridestream.adapters.*",
    "ridestream.application.spark_metrics",
]
disallow_untyped_defs = false
warn_return_any = false

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
    "UP",
    "ARG",
    "SIM",
    "TCH",
]
ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"src/ridestream/**/*.py" = ["TCH001", "TCH002", "TCH003"]
"tests/**/*.py" = ["TCH001", "TCH002", "TCH003"]

[tool.ruff.lint.isort]
known-first-party = ["ridestream"]
known-third-party = ["pydantic", "kafka", "boto3"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
py_version = 312
known_first_party = ["ridestream"]

[tool.bandit]
exclude_dirs = ["tests", "venv"]
skips = ["B101"]

[tool.detect-secrets]
baseline = ".secrets.baseline"
version = 1.4
disable_plugin_default = true
enable_plugins = [
    "detect_secrets.plugins.basic.BasicAuthDetector",
    "detect_secrets.plugins.aws.AWSKeyDetector",
]

[tool.coverage.run]
source = ["src/ridestream"]
omit = [
    "*/site-packages/*",
    "*/distutils/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
precision = 2

[dependency-groups]
dev = [
    "moto>=5.1.22",
    "mypy>=1.19.1",
    "pyparsing>=3.3.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
]
